move search to its own file
