+++ src/components/molecules/SearchUsers/index.ts
@@ -1,0 +1,3 @@
+import SearchUsers from './SearchUsers';
+
+export default SearchUsers;
