+++ src/components/molecules/index.ts
@@ -1,5 +1,6 @@
 import RecentUsersRow from './RecentUsersRow';
 import RecentUsersTable from './RecentUsersTable';
+import SearchUsers from './SearchUsers';
 import SelectedUsersTransfersTable from './SelectedUsersTransfersTable';
 import SelectedUserTransferRow from './SelectedUserTransferRow';
 import UserFlagsCard from './UserFlagsCard';
@@ -8,6 +9,7 @@ import UserProfileSummary from './UserProfileSummary';
 export {
   RecentUsersRow,
   RecentUsersTable,
+  SearchUsers,
   SelectedUsersTransfersTable,
   SelectedUserTransferRow,
   UserFlagsCard,
